% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkrot-package.R
\docType{package}
\name{linkrot-package}
\alias{linkrot}
\alias{linkrot-package}
\title{linkrot: Check Webpages For Linkrot}
\description{
Find and assess URLs from a webpage to see if they're still active.
   External references are broken when websites and pages disappear or move; a
   process known as 'link rot'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/linkrot/}
  \item Report bugs at \url{https://github.com/matt-dray/linkrot/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
