% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{lr_check}
\alias{lr_check}
\title{Assess a Webpage for Linkrot}
\usage{
lr_check(page)
}
\arguments{
\item{page}{Character. A valid https URL to a webpage that contains links you
want to check for linkrot.}
}
\value{
A data.frame.
}
\description{
Fetch the (https) URLs from a webpage and determine if they're still active
by checking for a response (i.e. a response code of 200).
}
\examples{
lr_check_page("https://www.rostrum.blog/2021/05/22/mission-across-iow/")
}
