% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{lr_check}
\alias{lr_check}
\title{Assess a Webpage for Linkrot}
\usage{
lr_check(page)
}
\arguments{
\item{page}{Character. A valid URL to a webpage that contains links you
want to check for linkrot. Must start 'http://www.' or 'https://www.'.}
}
\value{
A data.frame with 5 columns (page, link_url, link_text, response_code
and response_category, success) and as many rows per URLs found on the
page provided by the user.
}
\description{
Fetch web links from a given page and check for a response.
}
\examples{
\dontrun{ lr_check("https://www.rostrum.blog/") }
}
